backend/database/supabase_client.py
import os
from functools import lru_cache

from dotenv import load_dotenv
from supabase import Client, create_client

load_dotenv()

DEMO_MODE_MSG = "Supabase not configured. Add SUPABASE_URL and SUPABASE_SERVICE_KEY in backend/.env."


def supabase_configured() -> bool:
    return bool(os.getenv("SUPABASE_URL") and os.getenv("SUPABASE_SERVICE_KEY"))


@lru_cache
def get_supabase_client() -> Client:
    url = os.getenv("SUPABASE_URL")
    key = os.getenv("SUPABASE_SERVICE_KEY")
    if not url or not key:
        raise ValueError(DEMO_MODE_MSG)
    return create_client(url, key)
