backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
anthropic
pydantic
pydantic-settings
httpx
twilio
